fix(test/drivers): avoid using invalid pid numbers
In some systems the maximum value of pid is limited to a low
value such as e.g. 32768. Some of the driver tests are
currently using values that exceed this limit in certain systems.
This leads to failing driver tests due to an error in properly
setting up the test case.

Fit the value of the pid within the boudaries by reading the
system's maximum allowed pid value from
`/proc/sys/kernel/pid_max` and running modulos operator.

Signed-off-by: Tero Kauppinen <tero.kauppinen@est.tech>