new(sinsp): introduce plugin_tables class
This class holds all data related to accessing plugin-defined fields
and tables from libsinsp itself. None of this is the responsibility
of e.g. the thread_manager or individual threadinfos, but it has to live
somewhere. Rather than put it in the `sinsp` kitchen sink, move it all
to a separate class.

Note: at this point we still keep the accessors available through
thread_manager/threadinfo, they will be cleaned up later.

Co-Authored-By: Iacopo Rozzo <iacopo@sysdig.com>
Signed-off-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>