cleanup(sinsp)!: get_thread_ref now always does a proc lookup
All the remaining uses of get_thread_ref want a proc lookup so remove
the now redundant parameter.

The method is renamed to get_thread to force a compile error when
updating to a newer libs version, due to the two bool parameters
with defaults that could be confused and end up with different
semantics (the value of query_os_if_not_found would now become
the value of lookup_only).

Signed-off-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>