test(drivers/syscall_exit/execveat_x): fix `comm` assertion
`SyscallExit.execveatX_execve_exit_comm_equal_to_fd` test asserts the
`comm` parameter value among the others. For kernel versions lower
than 6.14, if the `AT_EMPTY_PATH` flag is specified while invoking
execveat, the comm value is expected to be set to the dirfd numeric
value. Starting from 6.14 (
https://github.com/torvalds/linux/commit/543841d1806029889c2f69f040e88b247aba8e22
), this strange behaviour has been fixed, and the exact same execveat
invocation results in the comm value to be correctly set to the
dentry's filename value. For this reason, account for both scenarios
while testing for the `comm` parameter to match the expectation.

Signed-off-by: Leonardo Di Giovanna <leonardodigiovanna1@gmail.com>