new(libsinsp): use plain type_id instead of full typeinfo
`type_id_of` is a constexpr variant of typeinfo::of which only
returns the type_id, not the whole typeinfo object.

Switch to using type_ids directly where possible, to avoid
the (mental) overhead of a dedicated class, and also to open
the possibility of using a different type for the underlying
field (typeinfo::of<FieldType> does not need to be the same
as the type we want to expose via the table API).

Signed-off-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>