fix(driver/bpf): fix sys_open_e on aarch64
On RHEL8 with clang-20, the verifier complains about unbounded values
for R2 (the read size).

After lots of (fruitless) experimenting with argument constraining, it
became apparent how the current implementation in the case of ia32
syscall will actually read 32 bits into the 64-bit value to be returned.
For some reason this makes clang 20 generate code with lots of spilling
just to preserve the other 32 bits (or so it seems).

Just use a different (scoped) variable of the appropriate type for each
of the two execution branches, which is what _READ_USER() actually does.

Signed-off-by: Gerlando Falauto <gerlando.falauto@sysdig.com>