fix(driver/bpf): fix sys_poll_e with clang20
clang20 tends to ignore (or optimize away) the checks we're performing
on read_size. The trick that seems to be working here is to first
perform our sanity check, then (uselessly) enforce data validity
by performing the AND operation, and finally (uselessly) create a
new exit path.

Notice how it's also important to move the reading of val at
the very beginning and away from read_size arithmetics.

Signed-off-by: Gerlando Falauto <gerlando.falauto@sysdig.com>