ci: install `bpftool` from released package if available
Install `bpftool` from released package if available. Currently, this
is the case for `amd64` and `arm64`, while it is not for `s390x` and
`ppc64le`. In these latter cases, install build it from sources and
install it. Keep using `bpftool` provided by the package manager on
`test-drivers-ppc64le` for the moment, until the self-hosted runner is
replaced.

Signed-off-by: Leonardo Di Giovanna <leonardodigiovanna1@gmail.com>