ci: add `install-cmake` composite action
Add `install-cmake` composite action, installing a specific version of
`cmake` from released packages. Currently, it's only used on `amd64`
and `arm64` runners. In the future, also use the action on `s390x` and
`ppc64le` runners if `cmake` starts to provide released packages for
these architectures. Moreover, it is not used on
`build-libs-linux-amd64-static`, as this requires a `cmake`
dynamically linked with `musl` in place of `glibc`.

Signed-off-by: Leonardo Di Giovanna <leonardodigiovanna1@gmail.com>