chore(userspace/libsinsp): remove unused `update-cri-proto` file
This file used to update the CRI protobuf dependency used by the old
container engines support. The support is now moved into a separate
container plugin, and interaction between the plugin and CRI is
handled through the Golang dependency system. This means that it is
possible to remove this file, since it is not used anymore by any
component.

Signed-off-by: Leonardo Di Giovanna <leonardodigiovanna1@gmail.com>